Refactored Component namings and added a burning loading spinner for web play back
